requirements.txt
langchain
langchain-community
langchain-groq
chromadb
pandas
streamlit
python-dotenv
sentence-transformers
langchain_huggingface